requirements.txt
# Data Processing
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.2.2
lightgbm --only-binary=:all:
statsmodels==0.14.0
torch==2.0.1

# API and Web
requests==2.31.0
beautifulsoup4==4.12.2
PyJWT==2.7.0
cryptography==41.0.1

# Data Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
networkx==3.1
PyYAML==6.0
joblib==1.2.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development Tools
black==23.3.0
flake8==6.0.0
mypy==1.3.0
jupyter==1.0.0
notebook==6.5.4

# Optional Dependencies
tqdm==4.65.0  # Progress bars
pyarrow==12.0.1  # Fast parquet support
plotly==5.15.0  # Interactive visualizations
streamlit>=1.28.0  # Dashboard framework
ipykernel==6.23.1  # Jupyter kernel